Moved tap actions to separate file
