+++ keyboards/40percentclub/gherkin/keymaps/knaruo/rules.mk
@@ -1,1 +1,3 @@
 TAP_DANCE_ENABLE = yes
+
+SRC += custom_tap.c
